[BOLT] build Argobots in the CMake working directory
Previously, the built-in Argobots was compiled in the Argobots source directory
(e.g., bolt/external/argobots), which messes up the source directory.  This
patch changes the build directory for the built-in Argobots so that it is
configured and build in CMAKE_CURRENT_BINARY_DIR/external/argobots
(e.g., bolt/build/external/argobots).

Reported-by: Michael Kruse <bolt@meinersbur.de>
Signed-off-by: Halim Amer <aamer@anl.gov>