[BOLT] temporarily install Argobots in the CMake working directory
Previously, the built-in Argobots was directly installed to the install path
(i.e., CMAKE_INSTALL_PREFIX). This behavior is not favored since actually
installs objects without user's explicit installation (e.g., make install).
This patch changes this behavior. The built-in Argobots is first installed to
the CMake working directory. It is copied to the CMake's installation directory
at installation time.

Because of this change, pkgconfig of the built-in Argobots gets incorrect, while
LLVM and it subprojects should use llvm-config.  This issue should be fixed if
LLVM adopts pkgconfig in the future.

Reported-by: Michael Kruse <bolt@meinersbur.de>
Signed-off-by: Halim Amer <aamer@anl.gov>