[OpenMP][Libomptarget] Fix check-libomptarget
The check-libomptarget fails when building with LLVM_ENABLE_PROJECTS. This is because test configuration misses the path to libomp.so and libLLVMSupport.so when time profiling is enabled (both libraries have the same path when building). This patch add the path to the configuration.

Reviewed By: vzakhari

Differential Revision: https://reviews.llvm.org/D95376

cherry-pick: 1e59c1a89851534f162d5ada9fb61796710e4de5
https://github.com/llvm/llvm-project/commit/1e59c1a89851534f162d5ada9fb61796710e4de5