[OpenMP] Fix misleading warning for OMP_PLACES
When OMP_PLACES contains an invalid value, the warning informs the user
that the fallback is OMP_PLACES=threads, but the actual internal setting
is OMP_PLACES=cores and is detected as such with KMP_SETTINGS=1.
This patch informs the user that OMP_PLACES=cores is being used instead
of OMP_PLACES=threads.

Differential Revision: https://reviews.llvm.org/D95170

cherry-pick: 8e6713436462daa43d84f8450d07a54f7657c31b
https://github.com/llvm/llvm-project/commit/8e6713436462daa43d84f8450d07a54f7657c31b