[OpenMP] Add environment variable to force monotonic dynamic scheduling
This patch introduces a new environment variable to force monotonic
behavior for users that absolutely need it.  This is in anticipation
of 5.0 change that uses non-monotonic behavior for dynamic scheduling
by default. Fixes for that and the actual switch are coming soon.

Differential Revision: https://reviews.llvm.org/D95263

cherry-pick: 67773681c0ef2a778a3c34d23b365899cd33e0c3
https://github.com/llvm/llvm-project/commit/67773681c0ef2a778a3c34d23b365899cd33e0c3