[OpenMP] Disable tests if FileCheck is not available in in-tree building
FileCheck is required for OpenMP tests. The current detection can fail
if building OpenMP in-tree when user sets `LLVM_INSTALL_TOOLCHAIN_ONLY=ON`. As a
result, CMake will raise an error and the compilation will be broken. This patch
fixed the issue. When `FileCheck` is not a target, tests will just be skipped.

Reviewed By: jdoerfert, JonChesterfield

Differential Revision: https://reviews.llvm.org/D95689

cherry-pick: f0129cc35ebd4607c86f1f90cb664c86ffdb4a5c
https://github.com/llvm/llvm-project/commit/f0129cc35ebd4607c86f1f90cb664c86ffdb4a5c