[OpenMP] Add lower and upper bound in num_teams clause
This patch adds lower-bound and upper-bound to num_teams clause
according to OpenMP 5.1 specification. The initial number of teams
created is implementation defined, but it will be greater than or
equal to lower-bound and less than or equal to upper-bound. If
num_teams clause is not specified, the number of teams created is
implementation defined, but it will be greater or equal to 1.

Differential Revision: https://reviews.llvm.org/D95820

cherry-pick: 4692bb4a8a6f78ee166ac8d2b1ec9082bd6e2be5
https://github.com/llvm/llvm-project/commit/4692bb4a8a6f78ee166ac8d2b1ec9082bd6e2be5