[libomptarget] Build amdgcn devicertl by default
[libomptarget] Build amdgcn devicertl by default

The cmake for this looks for an llvm install and does the right thing when
building as part of enable_runtimes. It will probably do the right thing
in other settings - at least, it won't try to build this with gcc.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D98658

cherry-pick: e23f3502d9a2b1e9abd445130a22cc00f6464da8
https://github.com/llvm/llvm-project/commit/e23f3502d9a2b1e9abd445130a22cc00f6464da8