[OpenMP] Reset async stream properly upon failure
Summary:
If the call to `synchronize` fails, it will currently block the stream indefinitely if execution is continued from this point. Additionally, if the program exits it will trigger an assertion on the non-null value of the async queue and prevent the runtime from printing debugging information.

Reviewers: jdoerfert

Differential Revision: https://reviews.llvm.org/D99443

cherry-pick: 16064e71e934547ca68b3d8f4bd71e9a2b7f4248
https://github.com/llvm/llvm-project/commit/16064e71e934547ca68b3d8f4bd71e9a2b7f4248