[OpenMP] Fix incorrect KMP_STRLEN() macro
The second argument to the strnlen_s(str, size) function should be
sizeof(str) when str is a true array of characters with known size
(instead of just a char*). Use type traits to determine if first
parameter is a character array and use the correct size based on that
trait.

Differential Revision: https://reviews.llvm.org/D98209

cherry-pick: 2aebb7cb3c88b1446515563653c821e8165b3aaf
https://github.com/llvm/llvm-project/commit/2aebb7cb3c88b1446515563653c821e8165b3aaf