[OpenMP] Replace global InfoLevel with a reference to an internal one.
Summary:
This patch improves the implementation of D100774 by replacing the global
variable introduced with a function that returns a reference to an internal
one. This removes the need to define the variable in every plugin that uses it.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D101102

cherry-pick: 59b68490122ae6ef92b1ebe45e8a5f2f7d88a401
https://github.com/llvm/llvm-project/commit/59b68490122ae6ef92b1ebe45e8a5f2f7d88a401