[libomptarget][amdgpu][nfc] Expand errorcheck macros
[libomptarget][amdgpu][nfc] Expand errorcheck macros

These macros expand to continue, which is confusing, or exit,
which is incompatible with continuing execution on offloading fail.

Expanding the macros in place makes the code look untidy but the
control flow obvious and amenable to improving. In particular, exit
becomes easier to eliminate.

Reviewed By: pdhaliwal

Differential Revision: https://reviews.llvm.org/D102230

cherry-pick: 9934571eab9c6b3be22c4c8857bd3f4280b77843
https://github.com/llvm/llvm-project/commit/9934571eab9c6b3be22c4c8857bd3f4280b77843