[libomptarget][amdgpu] Convert an assert to print and offload_fail
[libomptarget][amdgpu] Convert an assert to print and offload_fail

The kernel launched is supposed to be present in the binary, but a not yet
diagnosed bug means it is missing for some of the qmcpack test cases. Changing
from assert to print and offload_fail should help diagnose that and similar bugs.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D102378

cherry-pick: b049870d3b47a93b0c53f3ad69b11c4731b39d7f
https://github.com/llvm/llvm-project/commit/b049870d3b47a93b0c53f3ad69b11c4731b39d7f