[AMDGPU][Libomptarget] Remove g_atmi_machine global
Turns out the only purpose of this class was verify if device ID
was in range or not which could be done easily by using g_atl_machine.

Still getting rid of g_atl_machine is pending which would be done in
a later patch.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D103443

cherry-pick: fb113264a8da1d8e456e64f99920fac6b4fae548
https://github.com/llvm/llvm-project/commit/fb113264a8da1d8e456e64f99920fac6b4fae548